Add ExitCode<num> as test result check option
In addition to Pass and Fail expected outcomes where the testing harness
looks at the messages printed by a given test, ExitCode## can be given
to ignore output and check for a specific exit code from a test.  If the
exit code matches (e.g.  ExitCode0 was given and the test case exited
with status of 0) the test passes; otherwise it is flagged as a failure.

This extension to the testing harness is needed in order to test for
correct behavior of the shmem_global_exit function.

Signed-off-by: James Dinan <james.dinan@intel.com>